Split Plugins Into two entrypoint's (#36)
* chore: configure project's for github Releases

* fix: place conventional commit's in the version object

* fix: remove remark-html-directives-integration package

* fix: configure nx to skip lockfile update

* chore: test out pnpm update lock file
 it worked

* chore(release): publish

- project: utilities 4.12.0

- project: asciidoc 0.3.0

- project: flow 2.11.0

* fix: configure nx to skip lockfile update using generatorOptions
Nx can't use pnpm to update the lockfile this normally doesn't matter
I also allowed only packages to be published

* refactor: make package only compatable with 5.0.0 and up
From now on asciidoc can only be compatable with Astro 5.
BREAKING CHANGE: package is complete

* fix: remove flow package
I no longer plan to release version's of this library
Right now it's just in the way

* refactor: go back to manual commit's for now
There is no point in conventional commits now. The wrong files are due to the creation of different tag names. Nx doesn't allow you to change what tags are used for which project

* feat: create nu script for changing json files before release

* refactor: add nu script for refining dist json files to the pre version command

* fix: split plugin file into two different files
I ran into a bug where the tailwind plugin was supposed to be loaded but it couldn't due to the fact that it wasn't downloaded.
If the user uses the tailwind plugin then they should only use the tailwind file
The test file only tested the unocss logic so it's now changed to unocss.spec

* refactor: make the unocss and tailwind files into separate entry points
The user has to only use the file that uses the tool they want.
Esbuild is also configured to create the new entrypoints
BREAKING CHANGE: The plugins entrypoint is deleted

* refactor:adjust unocss to use the unocss entrypoint

* build: configure nx to use conventional commits when publishing a new package

* refactor: changed readme to account for changes